fix(init_spec): Clean up init_spec tests
I cleaned up the init_spec tests so that they now mostly pass.
Previously none of them passed, now most of them do.
The ones that do not are possibly actual bugs in the code.

The tests were previously separated into contexts based on OS, but did
not have if-statements preventing tests for one OS family running on
another family. This commit deduplicates the tests into a single set,
except for a few that use conditionals to prevent execution on
inappropriate operating systems.

This commit introduces the use of Ruby variables in testing, to
dynamically determine what the test results need to be based on the OS
being tested.